error handling for candidate and manager

+++ src/main/java/com/example/talent_api/service/CandidateService.java
@@ -42,8 +42,12 @@ public class CandidateService {
 
     }
 
-    public void deleteCandidate(Long id) {
-        candidateRepository.deleteById(id);
+    public boolean deleteCandidate(Long id) {
+        if(candidateRepository.existsById(id)){
+            candidateRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     public List<Candidate> searchCandidates (String searchTerm){
